refined the entire layout to accomodate the 3 scenarios (admin, dashboard and home pages)
