+++ app/(main)/layout.tsx
@@ -1,0 +1,17 @@
+import type React from "react";
+import { Header } from "@/components/layout/Header";
+import { Footer } from "@/components/layout/Footer";
+
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-gray-900 via-black to-gray-900">
+      <Header />
+      <main className="pt-18 md:pt-20">{children}</main>
+      <Footer />
+    </div>
+  );
+}
