--- app/packages/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return null
-}
